+++ src/Digdir.Domain.Dialogporten.Application/Common/Behaviours/FeatureMetric/FeatureMetricBehaviour.cs
@@ -24,16 +24,19 @@ internal sealed class FeatureMetricBehaviour<TRequest, TResponse>(
         CancellationToken cancellationToken)
     {
         var principal = _user.GetPrincipal();
-        principal.TryGetOrganizationShortName(out var performingOrg);
+        principal.TryGetConsumerOrgNumber(out var callerOrgNr);
+
         var hasAdminScope = principal.HasScope(AuthorizationScope.ServiceOwnerAdminScope);
         var resource = await _featureMetricServiceResourceResolver.Resolve(request, cancellationToken);
+
         _featureMetricRecorder.Record(new(
             FeatureName: typeof(TRequest).FullName!,
             HasAdminScope: hasAdminScope,
             Environment: hostEnvironment?.EnvironmentName,
-            PerformerOrg: performingOrg,
-            OwnerOrg: resource?.OwnOrgShortName,
+            CallerOrg: callerOrgNr,
+            OwnerOrg: resource?.OwnerOrgNumber,
             ServiceResource: resource?.ResourceId));
+
         return await next(cancellationToken);
     }
 }