+++ tests/k6/tests/enduser/all-tests.js
@@ -1,4 +1,5 @@
 // This file is generated, see "scripts" directory
+import { default as DialogGetAcceptLanguage } from './DialogGetAcceptLanguage.js';
 import { default as dialogBulkSystemLabels } from './dialogBulkSystemLabels.js';
 import { default as dialogDetails } from './dialogDetails.js';
 import { default as dialogSearch } from './dialogSearch.js';
@@ -8,6 +9,7 @@ import { default as dialogVisibleFrom } from './dialogVisibleFrom.js';
 import { default as parties } from './parties.js';
 
 export default function() {
+  DialogGetAcceptLanguage();
   dialogBulkSystemLabels();
   dialogDetails();
   dialogSearch();